requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-jose[cryptography]
jinja2 
python-multipart